File restructuring for the app
