+++ src/app.module.ts
@@ -4,10 +4,10 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
 import { MongooseModule } from '@nestjs/mongoose';
-import { FirebaseModule } from './firebase/firebase.module';
-import { UserModule } from './user/user.module';
-import { AssignmentModule } from './assignment/assignment.module';
-import { ChatModule } from './chat/chat.module';
+import { FirebaseModule } from './modules/firebase/firebase.module';
+import { UserModule } from './modules/user/user.module';
+import { AssignmentModule } from './modules/assignment/assignment.module';
+import { ChatModule } from './modules/chat/chat.module';
 import * as path from 'path';
 import { CourseModule } from './modules/course/course.module';
