+++ src/modules/assignment/assignment.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AssignmentController } from './assignment.controller';
 import { AssignmentService } from './assignment.service';
 import { MongooseModule } from '@nestjs/mongoose';
-import { AssignmentSchema } from '../schemas/assignment.schema';
+import { AssignmentSchema } from '../../schemas/assignment.schema';
 @Module({
   imports: [
     MongooseModule.forFeature([
