+++ src/modules/chat/chat.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ChatController } from './chat.controller';
 import { ChatService } from './chat.service';
 import { MongooseModule } from '@nestjs/mongoose';
-import { ChatSchema } from '../schemas/chat.schema';
+import { ChatSchema } from '../../schemas/chat.schema';
 @Module({
   imports: [MongooseModule.forFeature([{ name: 'Chat', schema: ChatSchema }])],
   controllers: [ChatController],
